upload

+++ controllers/menu.js
@@ -8,7 +8,13 @@ const Op = Sequelize.Op;
 
 router.get('/menus', async (req, res) => {
     try{
-        const menuList = await Menu.findAll();
+        const menuList = await Menu.findAll({
+            where:{
+                [Op.or]:[
+                    {name:{[Op.like]:"%"+req.query.kw+"%"}},
+                ]
+            },
+        });
         res.status(200).json(menuList);
     }catch(e){
         res.status(404).json(e);
@@ -32,9 +38,17 @@ router.post('/menu', sh.uploadImgFunc().single('icon'), async (req, res) => {
 
 router.delete('/menu', async (req, res) => {
     try{
+        const findItem = await Menu.findOne({where: {id: req.body.id}});
+        if(!findItem){ res.status(404).json('not foud'); return;}
+        try {
+            sh.deleteImgFunc(findItem.icon);
+          } catch (err) {}
+
         const removeModel = await Menu.destroy({
             where:{id: req.query.id}
         });
+
+        res.status(204).json(removeModel);
     }catch(e){
         res.status(404).json(e);
     }
@@ -57,7 +71,5 @@ router.put('/menu', sh.uploadImgFunc().single('icon'), async (req, res) => {
         res.status(404).json(e);
     }
 });
-
-
 
 module.exports = router;
